Created a module for the game struct
